requirements.txt
GitPython>=3.1.27
requests>=2.26.0
requests_cache>=0.9.5
